+++ app/src/main/java/com/wireguard/android/backend/WireGuardUiService.java
@@ -1,6 +1,5 @@
 package com.wireguard.android.backend;
 
-import android.Manifest;
 import android.app.Notification;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
@@ -11,11 +10,11 @@ import android.os.Binder;
 import android.os.Build;
 import android.os.CountDownTimer;
 import android.os.IBinder;
+import android.util.Log;
+
 import androidx.annotation.Nullable;
 import androidx.core.app.NotificationCompat;
 import androidx.localbroadcastmanager.content.LocalBroadcastManager;
-
-import android.util.Log;
 
 import net.ivpn.client.IVPNApplication;
 import net.ivpn.client.R;