+++ app/src/main/java/net/ivpn/client/ui/split/OnApplicationItemSelectionChangedListener.java
@@ -22,9 +22,12 @@ package net.ivpn.client.ui.split;
  along with the IVPN Android app. If not, see <https://www.gnu.org/licenses/>.
 */
 
-import net.ivpn.client.ui.split.data.ApplicationItem;
+import net.ivpn.client.ui.split.items.ApplicationItem;
 
 public interface OnApplicationItemSelectionChangedListener {
+
     void onApplicationItemSelectionChanged(ApplicationItem applicationItem, boolean isSelected);
+
     void onItemsSelectionStateChanged(boolean isAllItemSelected);
+
 }