+++ app/src/main/java/net/ivpn/client/v2/MainActivity.kt
@@ -31,6 +31,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.navigation.NavController
 import androidx.navigation.NavDestination
 import androidx.navigation.findNavController
+import androidx.navigation.fragment.NavHostFragment
 import net.ivpn.client.R
 import net.ivpn.client.common.extension.setContentSecure
 
@@ -47,8 +48,9 @@ class MainActivity : AppCompatActivity(), NavController.OnDestinationChangedList
     }
 
     override fun onDestroy() {
-        super.onDestroy()
+        //super.onDestroy() should be called at the end of the method. It's important.
         findNavController(R.id.nav_host_fragment).removeOnDestinationChangedListener(this)
+        super.onDestroy()
     }
 
     override fun onDestinationChanged(controller: NavController, destination: NavDestination, arguments: Bundle?) {