+++ app/src/main/java/net/ivpn/client/v2/settings/SettingsFragment.kt
@@ -62,8 +62,7 @@ import org.slf4j.LoggerFactory
 import java.util.*
 import javax.inject.Inject
 
-class SettingsFragment : Fragment(), KillSwitchViewModel.KillSwitchNavigator,
-        AdvancedKillSwitchActionListener, OnNightModeChangedListener, ColorThemeViewModel.ColorThemeNavigator {
+class SettingsFragment : Fragment(), OnNightModeChangedListener, ColorThemeViewModel.ColorThemeNavigator {
 
     companion object {
         private val LOGGER = LoggerFactory.getLogger(SettingsFragment::class.java)
@@ -93,9 +92,6 @@ class SettingsFragment : Fragment(), KillSwitchViewModel.KillSwitchNavigator,
     lateinit var antiTracker: AntiTrackerViewModel
 
     @Inject
-    lateinit var killSwitch: KillSwitchViewModel
-
-    @Inject
     lateinit var logging: LoggingViewModel
 
     @Inject
@@ -111,7 +107,7 @@ class SettingsFragment : Fragment(), KillSwitchViewModel.KillSwitchNavigator,
             inflater: LayoutInflater,
             container: ViewGroup?,
             savedInstanceState: Bundle?
-    ): View? {
+    ): View {
         binding = DataBindingUtil.inflate(inflater, R.layout.fragment_settings, container, false)
         return binding.root
     }
@@ -122,23 +118,6 @@ class SettingsFragment : Fragment(), KillSwitchViewModel.KillSwitchNavigator,
         initViews()
     }
 
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-
-        if (resultCode != Activity.RESULT_OK) {
-            LOGGER.debug("onActivityResult: RESULT_CANCELED")
-            return
-        }
-        LOGGER.debug("onActivityResult: RESULT_OK")
-
-        when (requestCode) {
-            ServiceConstants.ENABLE_KILL_SWITCH -> {
-                LOGGER.debug("onActivityResult: ENABLE_KILL_SWITCH")
-                killSwitch.enable(true)
-            }
-        }
-    }
-
     override fun onResume() {
         super.onResume()
 
@@ -146,7 +125,6 @@ class SettingsFragment : Fragment(), KillSwitchViewModel.KillSwitchNavigator,
         multihop.onResume()
         startOnBoot.onResume()
         alwaysOnVPN.onResume()
-        killSwitch.onResume()
         updates.onResume()
         logging.onResume()
         colorTheme.onResume()
@@ -154,17 +132,11 @@ class SettingsFragment : Fragment(), KillSwitchViewModel.KillSwitchNavigator,
 
     override fun onStart() {
         super.onStart()
-        killSwitch.navigator = this
         activity?.let {
             if (it is MainActivity) {
                 it.setContentSecure(false)
             }
         }
-    }
-
-    override fun onPause() {
-        super.onPause()
-        killSwitch.navigator = null
     }
 
     private fun initViews() {
@@ -175,7 +147,6 @@ class SettingsFragment : Fragment(), KillSwitchViewModel.KillSwitchNavigator,
         binding.contentLayout.startOnBoot = startOnBoot
         binding.contentLayout.alwaysOnVPN = alwaysOnVPN
         binding.contentLayout.antiTracker = antiTracker
-        binding.contentLayout.killSwitch = killSwitch
         binding.contentLayout.updates = updates
         binding.contentLayout.logging = logging
         binding.contentLayout.colorTheme = colorTheme
@@ -271,6 +242,22 @@ class SettingsFragment : Fragment(), KillSwitchViewModel.KillSwitchNavigator,
 
             openCustomDNSScreen()
         }
+        binding.contentLayout.sectionOther.killswitchLayout.setOnClickListener {
+            if (!account.authenticated.get()) {
+                openLoginScreen()
+                return@setOnClickListener
+            }
+            if (!account.isActive.get()) {
+                openAddFundsScreen()
+                return@setOnClickListener
+            }
+            if (connect.isVpnActive()) {
+                notifyUser(R.string.snackbar_to_use_kill_switch_disconnect)
+                return@setOnClickListener
+            }
+
+            openKillSwitchScreen()
+        }
         binding.contentLayout.sectionAbout.termsOfServiceLayout.setOnClickListener {
             openTermsOfServiceScreen()
         }
@@ -292,6 +279,15 @@ class SettingsFragment : Fragment(), KillSwitchViewModel.KillSwitchNavigator,
                 openEntryServerScreen()
             }
         }
+        binding.contentLayout.sectionServer.entryRandomLayout.setOnClickListener {
+            if (!account.authenticated.get()) {
+                openLoginScreen()
+            } else if (!account.isActive.get()) {
+                openAddFundsScreen()
+            } else {
+                openEntryServerScreen()
+            }
+        }
         binding.contentLayout.sectionServer.fastestServerLayout.setOnClickListener {
             if (!account.authenticated.get()) {
                 openLoginScreen()
@@ -302,6 +298,15 @@ class SettingsFragment : Fragment(), KillSwitchViewModel.KillSwitchNavigator,
             }
         }
         binding.contentLayout.sectionServer.exitServerLayout.setOnClickListener {
+            if (!account.authenticated.get()) {
+                openLoginScreen()
+            } else if (!account.isActive.get()) {
+                openAddFundsScreen()
+            } else {
+                openExitServerScreen()
+            }
+        }
+        binding.contentLayout.sectionServer.exitRandomLayout.setOnClickListener {
             if (!account.authenticated.get()) {
                 openLoginScreen()
             } else if (!account.isActive.get()) {
@@ -350,6 +355,11 @@ class SettingsFragment : Fragment(), KillSwitchViewModel.KillSwitchNavigator,
 
     private fun openCustomDNSScreen() {
         val action = SettingsFragmentDirections.actionSettingsFragmentToCustomDNSFragment()
+        navigate(action)
+    }
+
+    private fun openKillSwitchScreen() {
+        val action = SettingsFragmentDirections.actionSettingsFragmentToKillSwitchFragment()
         navigate(action)
     }
 
@@ -413,63 +423,8 @@ class SettingsFragment : Fragment(), KillSwitchViewModel.KillSwitchNavigator,
         navigate(action)
     }
 
-    private fun checkVPNPermission(requestCode: Int) {
-        val intent: Intent?
-        intent = try {
-            VpnService.prepare(context)
-        } catch (exception: Exception) {
-            exception.printStackTrace()
-            DialogBuilder.createNotificationDialog(context, Dialogs.FIRMWARE_ERROR)
-            return
-        }
-        if (intent != null) {
-            try {
-                startActivityForResult(intent, requestCode)
-            } catch (exception: ActivityNotFoundException) {
-                LOGGER.info("startVpnFromIntent: intent != null, ActivityNotFoundException")
-            }
-        } else {
-            onActivityResult(requestCode, Activity.RESULT_OK, null)
-        }
-    }
-
     private fun notifyUser(msgId: Int) {
         ToastUtil.toast(context, msgId)
-    }
-
-    override fun subscribe() {
-    }
-
-    override fun authenticate() {
-    }
-
-    override fun tryEnableKillSwitch(state: Boolean, advancedKillSwitchState: Boolean) {
-        LOGGER.info("enableKillSwitch = $state isAdvancedKillSwitchDialogEnabled = $advancedKillSwitchState")
-        if (state) {
-            checkVPNPermission(ServiceConstants.ENABLE_KILL_SWITCH)
-            if (killSwitch.isAdvancedModeSupported && advancedKillSwitchState) {
-                DialogBuilder.createAdvancedKillSwitchDialog(context, this)
-            }
-        } else {
-            killSwitch.enable(false)
-        }
-    }
-
-    override fun enableAdvancedKillSwitchDialog(enable: Boolean) {
-        LOGGER.info("enableAdvancedKillSwitchDialog")
-        killSwitch.enableAdvancedKillSwitchDialog(enable)
-    }
-
-    @SuppressLint("InlinedApi")
-    override fun openDeviceSettings() {
-        LOGGER.info("openDeviceSettings")
-        if (killSwitch.isAdvancedModeSupported) {
-            try {
-                startActivity(Intent(Settings.ACTION_VPN_SETTINGS))
-            } catch (exception: ActivityNotFoundException) {
-                DialogBuilder.createNotificationDialog(context, Dialogs.NO_VPN_SETTINGS)
-            }
-        }
     }
 
     override fun onNightModeChanged(mode: NightMode?) {