+++ app/src/main/java/net/ivpn/client/v2/sync/SyncFragment.kt
@@ -35,8 +35,6 @@ import net.ivpn.client.databinding.FragmentSyncBinding
 import net.ivpn.client.ui.syncservers.SyncServersNavigator
 import net.ivpn.client.ui.syncservers.SyncServersViewModel
 import net.ivpn.client.v2.MainActivity
-import net.ivpn.client.v2.connect.ConnectFragment
-import net.ivpn.client.v2.login.LoginFragment
 import org.slf4j.LoggerFactory
 import javax.inject.Inject